Import upstream version 1.2.20

+++ kclangc.cc
@@ -517,7 +517,7 @@ char* kcdbstatus(KCDB* db) {
  * Get keys matching a prefix string.
  */
 int64_t kcdbmatchprefix(KCDB* db, const char* prefix, char** strary, int64_t max) {
-  _assert_(db && prefix && strary && max >= 0 && max <= MEMMAXSIZ);
+  _assert_(db && prefix && strary && max >= 0 && max <= (int64_t)MEMMAXSIZ);
   PolyDB* pdb = (PolyDB*)db;
   std::vector<std::string> strvec;
   if (pdb->match_prefix(std::string(prefix), &strvec, max) == -1) return -1;
@@ -540,7 +540,7 @@ int64_t kcdbmatchprefix(KCDB* db, const char* prefix, char** strary, int64_t max
  * Get keys matching a regular expression string.
  */
 int64_t kcdbmatchregex(KCDB* db, const char* regex, char** strary, int64_t max) {
-  _assert_(db && regex && strary && max >= 0 && max <= MEMMAXSIZ);
+  _assert_(db && regex && strary && max >= 0 && max <= (int64_t)MEMMAXSIZ);
   PolyDB* pdb = (PolyDB*)db;
   std::vector<std::string> strvec;
   if (pdb->match_regex(std::string(regex), &strvec, max) == -1) return -1;
